Implement transmission fixity for external binaries. Add tests for transmission fixity, updating, and deleting external binaries.

+++ fcrepo-persistence-ocfl/src/main/java/org/fcrepo/persistence/ocfl/impl/DeleteResourcePersister.java
@@ -91,7 +91,9 @@ class DeleteResourcePersister extends AbstractPersister {
     private void deletePath(final String path, final OcflObjectSession session, final ResourceHeadersImpl headers,
                             final String user, final Instant deleteTime, final String headerPath)
         throws PersistentStorageException {
-        session.delete(path);
+        if (path != null) {
+            session.delete(path);
+        }
         headers.setDeleted(true);
         ResourceHeaderUtils.touchModificationHeaders(headers, user, deleteTime);
         writeHeaders(session, headers, headerPath);
